+++ faassupervisor/storage/config.py
@@ -14,9 +14,30 @@
 """Class to parse, store and manage storage information."""
 from faassupervisor.utils import ConfigUtils, FileUtils
 from faassupervisor.exceptions import NoOutputStorageProviderDefinedWarning, \
-    NoStorageProviderDefinedWarning, StorageAuthError
+    NoStorageProviderDefinedWarning, StorageAuthError, InvalidStorageProviderError
+from faassupervisor.storage.providers.local import Local
+from faassupervisor.storage.providers.minio import Minio
+from faassupervisor.storage.providers.onedata import Onedata
+from faassupervisor.storage.providers.s3 import S3
 from faassupervisor.logger import get_logger
-from faassupervisor.storage.providers import create_provider
+
+
+def create_provider(storage_auth):
+    """Returns the storage provider needed
+    based on the authentication type defined.
+
+    If not storage auth provided, use local storage."""
+    if not storage_auth:
+        provider = Local(storage_auth)
+    elif storage_auth.type == 'MINIO':
+        provider = Minio(storage_auth)
+    elif storage_auth.type == 'ONEDATA':
+        provider = Onedata(storage_auth)
+    elif storage_auth.type == 'S3':
+        provider = S3(storage_auth)
+    else:
+        raise InvalidStorageProviderError(storage_type=storage_auth.type)
+    return provider
 
 
 class AuthData():
